fix frontend sentry setup
